requirements.txt
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
cffi==1.17.1
cryptography==45.0.6
numpy==2.3.2
PyAudio==0.2.14
pycparser==2.22
pynput==1.8.1
six==1.17.0
